Renames
- Rename WEX to Wolverine
- Fix GH CI
