--- QuantConnect.WEX.Tests/QuantConnect.WEX.Tests.csproj
@@ -1,45 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<Configuration Condition=" '$(Configuration)' == '' ">Release</Configuration>
-		<Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-		<TargetFramework>net6.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-
-		<IsPackable>false</IsPackable>
-		<Copyright>Copyright ©  2021</Copyright>
-		<TestProjectType>UnitTest</TestProjectType>
-		<OutputPath>bin\$(Configuration)\</OutputPath>
-		<Product>QuantConnect.WEX.Tests</Product>
-		<AssemblyName>QuantConnect.WEX.Tests</AssemblyName>
-		<RootNamespace>QuantConnect.WEX.Tests</RootNamespace>
-		<AssemblyTitle>QuantConnect.WEX.Tests</AssemblyTitle>
-		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
-		<PackageReference Include="NUnit" Version="3.13.3" />
-		<PackageReference Include="NUnit3TestAdapter" Version="4.2.1">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
-		<PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
-		<PackageReference Include="coverlet.collector" Version="3.1.2" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\..\Lean\Tests\QuantConnect.Tests.csproj" />
-		<ProjectReference Include="..\QuantConnect.WEX\QuantConnect.WEX.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<None Remove="config.json" />
-		<Content Include="config.json">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</Content>
-	</ItemGroup>
-
-</Project>
