--- QuantConnect.WEX.Tests/Usings.cs
@@ -1,17 +1,0 @@
-global using NUnit.Framework;
-global using System.Collections;
-global using QuantConnect.Util;
-global using QuantConnect.Orders;
-global using QuantConnect.Packets;
-global using QuantConnect.WEX.Fix;
-global using QuantConnect.WEX.Wex;
-global using QuantConnect.Logging;
-global using QuantConnect.Algorithm;
-global using QuantConnect.Interfaces;
-global using QuantConnect.WEX.Fix.Core;
-global using QuantConnect.Configuration;
-global using QuantConnect.Tests.Brokerages;
-global using Log = QuantConnect.Logging.Log;
-global using QuickFix;
-
-
