--- QuantConnect.WEX/Fix/LogFactory/QuickFixLogFactory.cs
@@ -1,36 +1,0 @@
-﻿/*
- * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
- * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-using QuickFix;
-using System.Collections.Concurrent;
-
-namespace QuantConnect.WEX.Fix.LogFactory
-{
-    public class QuickFixLogFactory : ILogFactory
-    {
-        private static readonly ConcurrentDictionary<SessionID, ILog> Loggers = new ConcurrentDictionary<SessionID, ILog>();
-        private readonly bool _logFixMessages;
-
-        public QuickFixLogFactory(bool logFixMessages)
-        {
-            _logFixMessages = logFixMessages;
-        }
-
-        public ILog Create(SessionID sessionId)
-        {
-            return Loggers.GetOrAdd(sessionId, s => new QuickFixLogger(s, _logFixMessages));
-        }
-    }
-}
