--- QuantConnect.WEX/Fix/Protocol/IFixOutboundBrokerageHandler.cs
@@ -1,45 +1,0 @@
-﻿/*
- * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
- * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-using QuantConnect.Orders;
-
-namespace QuantConnect.WEX.Fix.Protocol
-{
-    public interface IFixOutboundBrokerageHandler
-    {
-        bool RequestOpenOrders();
-
-        /// <summary>
-        /// Places an order.
-        /// </summary>
-        /// <param name="order">Order to submit</param>
-        /// <returns>Whether the request was sent</returns>
-        bool PlaceOrder(Order order);
-
-        /// <summary>
-        /// Updates an order.
-        /// </summary>
-        /// <param name="order">Order to update</param>
-        /// <returns>Whether the request was sent</returns>
-        bool UpdateOrder(Order order);
-
-        /// <summary>
-        /// Cancels an order.
-        /// </summary>
-        /// <param name="order">Order to cancel</param>
-        /// <returns>Whether the request was sent</returns>
-        bool CancelOrder(Order order);
-    }
-}
