--- QuantConnect.WEX/QuantConnect.WEX.csproj
@@ -1,41 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<Configuration Condition=" '$(Configuration)' == '' ">Release</Configuration>
-		<Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-		<TargetFramework>net6.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<Product>QuantConnect.WEX</Product>
-		<AssemblyName>QuantConnect.WEX</AssemblyName>
-		<RootNamespace>QuantConnect.WEX</RootNamespace>
-		<AssemblyTitle>QuantConnect.WEX</AssemblyTitle>
-		<OutputType>Library</OutputType>
-		<OutputPath>bin\$(Configuration)\</OutputPath>
-		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-		<GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-		<Description>QuantConnect LEAN WEX Brokerage: WEX Brokerage plugin for Lean</Description>
-	</PropertyGroup>
-	<PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-		<DebugType>full</DebugType>
-		<OutputPath>bin\Debug\</OutputPath>
-	</PropertyGroup>
-	<PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-		<DebugType>pdbonly</DebugType>
-		<OutputPath>bin\Release\</OutputPath>
-	</PropertyGroup>
-	<ItemGroup>
-		<Compile Include="..\..\Lean\Common\Properties\SharedAssemblyInfo.cs" Link="Properties\SharedAssemblyInfo.cs" />
-	</ItemGroup>
-	<ItemGroup>
-		<PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-		<PackageReference Include="QuantConnect.Brokerages" Version="2.5.*" />
-		<PackageReference Include="QuickFIXn.Core" Version="1.10.0" />
-		<PackageReference Include="QuickFIXn.FIX4.2" Version="1.10.0" />
-	</ItemGroup>
-	<ItemGroup>
-	  <None Update="Wex\WEX-FIX42.xml">
-	    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-	  </None>
-	</ItemGroup>
-</Project>
