+++ QuantConnect.Wolverine.Tests/Usings.cs
@@ -1,0 +1,16 @@
+global using NUnit.Framework;
+global using System.Collections;
+global using QuantConnect.Util;
+global using QuantConnect.Orders;
+global using QuantConnect.Packets;
+global using QuantConnect.Wolverine.Fix;
+global using QuantConnect.Logging;
+global using QuantConnect.Algorithm;
+global using QuantConnect.Interfaces;
+global using QuantConnect.Wolverine.Fix.Core;
+global using QuantConnect.Configuration;
+global using QuantConnect.Tests.Brokerages;
+global using Log = QuantConnect.Logging.Log;
+global using QuickFix;
+
+
