+++ QuantConnect.Wolverine.Tests/config.json
