+++ QuantConnect.Wolverine.sln
@@ -15,9 +15,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuantConnect.Tests", "..\Le
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuantConnect.ToolBox", "..\Lean\ToolBox\QuantConnect.ToolBox.csproj", "{88885A3A-5028-4EFB-9244-47FDB04725CE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuantConnect.WEX", "QuantConnect.WEX\QuantConnect.WEX.csproj", "{4E33C7FC-43FB-477B-B798-9ECAEA1025FB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuantConnect.Wolverine", "QuantConnect.Wolverine\QuantConnect.Wolverine.csproj", "{4E33C7FC-43FB-477B-B798-9ECAEA1025FB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuantConnect.WEX.Tests", "QuantConnect.WEX.Tests\QuantConnect.WEX.Tests.csproj", "{78FB3A52-FA73-4E1D-BB19-7F814DB73688}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuantConnect.Wolverine.Tests", "QuantConnect.Wolverine.Tests\QuantConnect.Wolverine.Tests.csproj", "{78FB3A52-FA73-4E1D-BB19-7F814DB73688}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
